Packet prices added

+++ app/Http/Controllers/Web/V1/Admin/System/Course/PacketCourseController.php
@@ -4,6 +4,7 @@
 namespace App\Http\Controllers\Web\V1\Admin\System\Course;
 
 
+use App\Exceptions\WebServiceErroredException;
 use App\Http\Controllers\Web\WebBaseController;
 use App\Models\Entities\Course\Course;
 use App\Models\Entities\Course\Packet;
@@ -24,8 +25,14 @@ class PacketCourseController extends WebBaseController
 
     public function connect($packet_id, $course_id)
     {
+        $packet = Packet::with('course')->findOrFail($packet_id);
+
+        if ($packet->course && !$packet->course->is_parent) {
+            throw new WebServiceErroredException("Это курс не является пакетным");
+        }
+
         $courses = PacketCourse::where('packet_id', $packet_id)->where('course_id', $course_id)->get();
-        if($courses->isEmpty()){
+        if ($courses->isEmpty()) {
             PacketCourse::create([
                 'packet_id' => $packet_id,
                 'course_id' => $course_id
@@ -36,6 +43,11 @@ class PacketCourseController extends WebBaseController
 
     public function disconnect($packet_id, $course_id)
     {
+        $packet = Packet::with('course')->findOrFail($packet_id);
+
+        if ($packet->course && !$packet->course->is_parent) {
+            throw new WebServiceErroredException("Это курс не является пакетным");
+        }
         PacketCourse::where('packet_id', $packet_id)->where('course_id', $course_id)->delete();
         return redirect()->back();
     }
